requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.9.5=pypi_0
aiohttp-retry=2.8.3=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.2=pyhd8ed1ab_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
apscheduler=3.10.4=pyhd8ed1ab_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
atk-1.0=2.38.0=h6b5321d_1
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
bcrypt=4.0.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
brotli=1.1.0=pypi_0
brotli-bin=1.0.9=h2bbff1b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.4=h56e8100_0
cairo=1.18.0=h1fef639_0
certifi=2024.7.4=pypi_0
cffi=1.16.0=py312h2bbff1b_1
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
cssselect2=0.7.0=pypi_0
debugpy=1.8.2=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.2.0=pypi_0
epoxy=1.5.10=h8d14728_1
executing=2.0.1=pypi_0
expat=2.6.2=hd77b12b_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
fastjsonschema=2.20.0=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_2
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.53.1=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
fribidi=1.0.10=h62dcd97_0
frozenlist=1.4.1=pypi_0
gdk-pixbuf=2.42.10=h84213e4_0
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
graphite2=1.3.14=hd77b12b_1
greenlet=3.0.3=py312h53d5487_0
gtk3=3.24.41=ha29fada_0
h11=0.14.0=pypi_0
harfbuzz=8.3.0=h7ab893a_0
hicolor-icon-theme=0.17=h57928b3_2
html5lib=1.1=pyhd3eb1b0_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
icu=73.2=h63175ca_0
idna=3.7=py312haa95532_0
importlib-metadata=8.0.0=pyha770c72_0
importlib_resources=6.4.0=pyhd8ed1ab_0
ipykernel=6.29.5=pypi_0
ipython=8.26.0=pypi_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
jpeg=9e=h2bbff1b_1
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.1=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.3=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libdeflate=1.17=h2bbff1b_1
libexpat=2.6.2=h63175ca_0
libffi=3.4.4=hd77b12b_1
libglib=2.78.4=ha17d25a_0
libiconv=1.17=hcfcfb64_2
libpng=1.6.39=h8cc25b3_0
libsqlite=3.46.0=h2466b09_0
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_2
libzlib=1.2.13=h2466b09_6
lz4-c=1.9.4=h2bbff1b_1
mako=1.3.5=pyhd8ed1ab_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=py312he70551f_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
notebook=7.2.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=2.0.0=pypi_0
openjpeg=2.4.0=h4afccc4_2
openssl=3.3.1=h2466b09_1
orjson=3.10.6=pypi_0
overrides=7.7.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
pango=1.50.14=h07c897b_2
parso=0.8.4=pypi_0
passlib=1.7.4=pypi_0
pcre2=10.42=h0ff8eda_1
pillow=10.4.0=pypi_0
pip=24.0=py312haa95532_0
pixman=0.43.4=h63175ca_0
platformdirs=4.2.2=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.47=pypi_0
psutil=6.0.0=pypi_0
psycopg2=2.9.9=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.22=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pydyf=0.11.0=pyhd8ed1ab_0
pygments=2.18.0=pypi_0
pyjwt=2.8.0=pypi_0
pyphen=0.15.0=pypi_0
python=3.12.3=h2628c8c_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-http-client=3.3.7=pypi_0
python-json-logger=2.0.7=pypi_0
python-multipart=0.0.9=pypi_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytz=2024.1=pyhd8ed1ab_0
pywin32=306=pypi_0
pywinpty=2.0.13=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
send2trash=1.8.3=pypi_0
sendgrid=6.11.0=pypi_0
setuptools=70.2.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.31=py312h4389bb4_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
starkbank-ecdsa=2.2.0=pypi_0
starlette=0.37.2=pypi_0
terminado=0.18.1=pypi_0
tinycss2=1.3.0=pyhd8ed1ab_0
tk=8.6.14=h0416ee5_0
tornado=6.4.1=pypi_0
traitlets=5.14.3=pypi_0
twilio=9.2.3=pypi_0
typer=0.12.3=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
tzlocal=5.2=py312h2e8e312_0
ucrt=10.0.22621.0=h57928b3_0
ujson=5.10.0=pypi_0
unicodedata2=15.1.0=py312h2bbff1b_0
uri-template=1.3.0=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.1=pypi_0
vc=14.2=h2eaa2aa_4
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
watchfiles=0.22.0=pypi_0
wcwidth=0.2.13=pypi_0
weasyprint=62.3=pyhd8ed1ab_0
webcolors=24.6.0=pypi_0
webencodings=0.5.1=py312haa95532_2
websocket-client=1.8.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py312haa95532_0
widgetsnbextension=4.0.11=pypi_0
xz=5.4.6=h8cc25b3_1
yarl=1.9.4=pypi_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h2466b09_6
zopfli=0.2.3=pypi_0
zstd=1.5.5=hd43e919_2
